// apps/api/src/generated/prisma/models/Reservation.ts
export type ReservationModel = runtime.Types.Result.DefaultSelection<Prisma.$ReservationPayload>
// ...
export type $ReservationPayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "Reservation"
  objects: {
    user: Prisma.$UserPayload<ExtArgs>
    event: Prisma.$EventPayload<ExtArgs>
  }
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: string
    userId: string
    eventId: string
    status: $Enums.ReservationStatus
    createdAt: Date
  }, ExtArgs["result"]["reservation"]>
  composites: {}
}