// apps/api/src/generated/prisma/models/Event.ts
export type EventModel = runtime.Types.Result.DefaultSelection<Prisma.$EventPayload>
// ...
export type $EventPayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "Event"
  objects: {
    reservations: Prisma.$ReservationPayload<ExtArgs>[]
  }
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: string
    title: string
    description: string
    dateTime: Date
    location: string
    maxCapacity: number
    status: $Enums.EventStatus
    createdAt: Date
  }, ExtArgs["result"]["event"]>
  composites: {}
}